requirements.txt
crewai==0.22.5

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.5

python-dotenv==1.0.0
youtube-transcript-api==0.6.2
yt-dlp==2024.12.6

streamlit==1.32.0